Implemented MinMax LLM HTTP service support.
Added the ability to communicate with the MinMax LLM HTTP service. This includes adding the corresponding URL and API key in the configuration, and implementing the sending and processing of HTTP POST requests in the code. This update mainly involves the expansion of the voice assistant configuration, as well as the detailed code implementation of the new HTTP service implementation